requirements.txt
black
pylint
numpy
pillow
mss
pygetwindow
opencv-python
matplotlib
pandas
stable-baselines3
pywin32
debugpy
pytesseract
tensorboard
playsound==1.2.2